requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.0.0
jinja2>=3.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
